methods, routes were explored
