+++ 02-express-tutorial/routes/people.js
@@ -1,0 +1,66 @@
+// people.js
+
+const express = require('express');
+const router = express.Router();
+let { people } = require('../data');
+
+
+router.get('/', (req, res) => {
+    res.status(200).json({success: true, data: people});
+})
+
+router.post('/', (req, res) => {
+    const { name } = req.body
+    if (!name) {
+        return res
+            .status(400)
+            .json({ success: false, msg: 'please provide name value' })
+    }
+    res.status(201).json({ success: true, person: name })
+})
+
+router.post('/postman', (req, res) => {
+    const { name } = req.body
+    if (!name) {
+        return res
+            .status(400)
+            .json({ success: false, msg: 'please provide name value' })
+    }
+    res.status(201).json({ success: true, data: [...people, name] })
+})
+
+router.put('/:id', (req, res) => {
+    const { id } = req.params
+    const { name } = req.body
+    console.log(id, name)
+    // res.send('hello')
+    const person = people.find((person) => person.id === Number(id))
+
+    if (!person) {
+        return res
+            .status(404)
+            .json({ success: false, msg: `no person with id; ${id}` })
+    }
+    const newPerson = people.map((person) => {
+        if (person.id === Number(id)) {
+            person.name = name
+        }
+        return person
+    })
+    return res.status(200).json({ success: true, data: newPerson })
+})
+
+router.delete('/:id', (req, res) => {
+    const person = people.find((person) => person.id === Number(req.params.id))
+    if (!person) {
+        return res
+            .status(404)
+            .json({ success: false, msg: `no person with id ${req.params.id}` })
+    }
+    const newPeople = people.filter(
+        (person) => person.id !== Number(req.params.id)
+    )
+    return res.status(200).json({ success: true, data: newPeople })
+})
+
+module.exports = router;
